Prevent warm_start same classes error
Using warm_start in conjunction with max_iter=1 may result in the
following error during a call to fit():

ValueError: warm_start can only be used where `y` has the same classes
as in the previous call to fit. Previously got [0 1], `y` has [0]

This is due to a certain peculiarity in fit() implementation.
partial_fit() is not affected by it, and, in fact, it is a superior
option in terms of code cleanliness anyway.